backend/requirements.txt
asgiref==3.7.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.2.0
cryptography==41.0.7
defusedxml==0.7.1
Django==4.2.5
django-allauth==0.59.0
django-rest-passwordreset==1.3.0
Django-Verify-Email==2.0.3
djangorestframework==3.14.0
gunicorn==21.2.0
idna==3.4
oauthlib==3.2.2
packaging==23.1
psycopg==3.1.12
psycopg-binary==3.1.12
pycparser==2.21
PyJWT==2.8.0
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
sqlparse==0.4.4
typing_extensions==4.8.0
urllib3==2.0.5
